[UHYU-171] refactor: 코드리뷰 반영 - 기본 브랜드 로고 추가 및 카테고리, 브랜드 관련 상수 파일 생성, 코드 정리
