+++ src/features/kakao-map/components/layout/steps/BrandSelectContent.tsx
@@ -38,7 +38,9 @@ const BrandSelectContent: FC<BrandSelectContentProps> = ({
    * 부모 컴포넌트에 선택 결과를 전달합니다.
    */
   const handleBrandClick = (brandName: string) => {
-    console.log('🏢 브랜드 선택:', brandName);
+    if (process.env.NODE_ENV === 'development') {
+      console.log('🏢 브랜드 선택:', brandName);
+    }
     onBrandSelect(brandName);
   };
