+++ src/features/kakao-map/constants/brands.ts
@@ -1,0 +1,10 @@
+export const BRAND_MAPPING: Record<string, string[]> = {
+  lifestyle: ['GS25', '펫생각', '셀로', '다락'],
+  food: ['스타벅스', '투썸플레이스', '이디야', '파리바게뜨'],
+  beauty: ['올리브영', '아리따움', '롭스', '왓슨스'],
+  shopping: ['이마트', '홈플러스', '롯데마트', '코스트코'],
+  culture: ['CGV', '롯데시네마', '메가박스', '교보문고'],
+  activity: ['피트니스클럽', '골프장', '볼링장', '당구장'],
+  education: ['해커스', '시원스쿨', '에듀윌', '종로학원'],
+  travel: ['하나투어', '모두투어', '노랑풍선', '온라인투어'],
+};
