+++ src/features/kakao-map/constants/categories.ts
@@ -1,0 +1,10 @@
+export const CATEGORIES = [
+  { key: 'lifestyle', name: '생활/편의', icon: '🏪' },
+  { key: 'food', name: '푸드', icon: '🍽️' },
+  { key: 'beauty', name: '뷰티/건강', icon: '💄' },
+  { key: 'shopping', name: '쇼핑', icon: '🛍️' },
+  { key: 'culture', name: '문화/여가', icon: '🎬' },
+  { key: 'activity', name: '액티비티', icon: '🏃' },
+  { key: 'education', name: '교육', icon: '📚' },
+  { key: 'travel', name: '여행/교통', icon: '✈️' },
+];
