+++ src/features/kakao-map/hooks/useFilterNavigation.ts
@@ -1,5 +1,6 @@
 import { useCallback } from 'react';
 import { useMapContext } from '../context/MapContext';
+import { BRAND_MAPPING } from '../constants/brands';
 
 export const useFilterNavigation = () => {
   const { state, actions } = useMapContext();
@@ -19,27 +20,22 @@ export const useFilterNavigation = () => {
   }, []);
 
   const getBrandsByCategory = useCallback((category: string): string[] => {
-    const brandMapping: Record<string, string[]> = {
-      lifestyle: ['GS25', '펫생각', '셀로', '다락'],
-      food: ['스타벅스', '투썸플레이스', '이디야', '파리바게뜨'],
-      beauty: ['올리브영', '아리따움', '롭스', '왓슨스'],
-      shopping: ['이마트', '홈플러스', '롯데마트', '코스트코'],
-      culture: ['CGV', '롯데시네마', '메가박스', '교보문고'],
-      activity: ['피트니스클럽', '골프장', '볼링장', '당구장'],
-      education: ['해커스', '시원스쿨', '에듀윌', '종로학원'],
-      travel: ['하나투어', '모두투어', '노랑풍선', '온라인투어'],
-    };
+    const brandMapping = BRAND_MAPPING;
     return brandMapping[category] || [];
   }, []);
 
   const handleShowFilter = useCallback(() => {
-    console.log('🔍 필터 설정 시작 - 카테고리 선택 화면으로 이동');
+    if (process.env.NODE_ENV === 'development') {
+      console.log('🔍 필터 설정 시작 - 카테고리 선택 화면으로 이동');
+    }
     actions.setBottomSheetStep('category');
   }, [actions]);
 
   const handleCategorySelect = useCallback(
     (categoryKey: string) => {
-      console.log('📂 카테고리 선택:', categoryKey);
+      if (process.env.NODE_ENV === 'development') {
+        console.log('📂 카테고리 선택:', categoryKey);
+      }
       actions.setSelectedCategory(categoryKey);
       actions.setBottomSheetStep('brand');
     },
@@ -48,7 +44,9 @@ export const useFilterNavigation = () => {
 
   const handleBrandSelect = useCallback(
     (brandName: string) => {
-      console.log('🏢 브랜드 선택:', brandName);
+      if (process.env.NODE_ENV === 'development') {
+        console.log('🏢 브랜드 선택:', brandName);
+      }
       actions.setSelectedBrand(brandName);
       actions.setBottomSheetStep('list');
       // TODO: 실제 필터링 로직 적용
@@ -57,7 +55,9 @@ export const useFilterNavigation = () => {
   );
 
   const handleBackToCategory = useCallback(() => {
-    console.log('⬅️ 카테고리 선택 화면으로 돌아가기');
+    if (process.env.NODE_ENV === 'development') {
+      console.log('⬅️ 카테고리 선택 화면으로 돌아가기');
+    }
     actions.setBottomSheetStep('category');
   }, [actions]);
